+++ qwt/src/qwt_plot_spectrocurve.cpp
@@ -243,7 +243,7 @@ void QwtPlotSpectroCurve::drawSeries( QPainter* painter,
         return;
 
     if ( to < 0 )
-        to = dataSize() - 1;
+        to = static_cast<int>( dataSize() - 1 );
 
     if ( from < 0 )
         from = 0;
